Attempt to transfer logic test logic into test/docker/
